+++ src/main/java/com/wenxin2/marioverse/blocks/PicketFenceBlock.java
@@ -11,10 +11,13 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import net.minecraft.ChatFormatting;
+import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.GlobalPos;
 import net.minecraft.core.particles.DustParticleOptions;
+import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
@@ -26,7 +29,9 @@ import net.minecraft.world.ItemInteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.DyeItem;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
@@ -171,6 +176,21 @@ public class PicketFenceBlock extends HorizontalDirectionalBlock implements Simp
         Direction target = (shape == StairsShape.OUTER_LEFT || shape == StairsShape.INNER_LEFT)
                 ? state.getValue(FACING).getCounterClockWise() : state.getValue(FACING);
         return this.rotateShape(Direction.NORTH, target, base);
+    }
+
+    @Override
+    public void appendHoverText(ItemStack stack, Item.TooltipContext tooltipContext, List<Component> list, TooltipFlag options) {
+        super.appendHoverText(stack, tooltipContext, list, options);
+
+        if (Screen.hasShiftDown()) {
+            list.add(Component.literal(""));
+
+            list.add(Component.translatable("block.marioverse.picket_fence.tooltip.instructions"));
+            list.add(Component.translatable("block.marioverse.picket_fence.tooltip.instructions.dye"));
+            list.add(Component.translatable("block.marioverse.picket_fence.tooltip.instructions.dye_one"));
+
+            list.add(Component.literal(""));
+        } else list.add(Component.translatable("block.marioverse.picket_fence.tooltip"));
     }
 
     @Nullable